products/requirements.txt
fastapi
uvicorn
mysql-connector-python
redis
PyJWT